enhance: Introduce user settings panel for profile management, theme, and view preferences, alongside updates to exam components and development tooling.
