+++ .husky/pre-commit
@@ -1,2 +1,2 @@
 
-pnpm lint-staged
+bun run lint-staged
